[mapleall]Internal commit msg:     [mapleall] CleanCode 7

+++ src/mapleall/maple_be/src/cg/aarch64/aarch64_memlayout.cpp
@@ -433,7 +433,7 @@ void AArch64MemLayout::LayoutStackFrame(int32 &structCopySize, int32 &maxParmSta
     /* we do need this as SP has to be aligned at a 16-bytes bounardy */
     segArgsStkPassed.SetSize(RoundUp(segArgsStkPassed.GetSize(), k8ByteSize + k8ByteSize));
   } else {
-    segArgsRegPassed.SetSize(RoundUp(segArgsRegPassed.GetSize(), GetPointerSize()));
+    segArgsRegPassed.SetSize(static_cast<uint32>(RoundUp(segArgsRegPassed.GetSize(), GetPointerSize())));
     segArgsStkPassed.SetSize(static_cast<uint32>(RoundUp(segArgsStkPassed.GetSize(), GetPointerSize() +
                                                                                      GetPointerSize())));
   }
